Update parse_apache_file.pl
################################################################################################################################################
The perl script uses the Parse::Apache::ServerStatus to fetch the apache statistics.
The script takes the hostname/ipaddress as the only argurement.if none is passed , "localhost" is used.
The script uses hostname to create the apache server-status url. http://hostname/server-status
Please make sure that server-status is enabled on server.